Yaci DevKit Integration

+++ src/main/java/com/bloxbean/intelliada/idea/nodeint/service/impl/AccountServiceImpl.java
@@ -35,7 +35,7 @@ public class AccountServiceImpl extends NodeBaseService implements CardanoAccoun
     }
 
     @Override
-    public Result<Long> getAdaBalance(String address) {
+    public Result<BigInteger> getAdaBalance(String address) {
 
         LedgerAccount account;
         AddressContent addressContent = null;
